add fixtures for unit tests
