+++ src/__fixtures__/ServiceFixture.tsx
@@ -1,0 +1,9 @@
+import { selectedServices } from '../redux/reducers/BerthReducers';
+import { SelectedServices, SelectedServicesProps } from '../types/services';
+
+export const createServices = (options?: object): SelectedServices => {
+  const defaultServices = selectedServices();
+  return defaultServices.merge({ ...options });
+};
+
+export const services = createServices();
